Convert the list of questions to a map
The previous variable questions has now been converted to a map with key value pairs of questions and answers. The questionText key takes the value of a string where questions to be asked is passed on. The answerText key takes a list of answers available for the given question. Finally, the Answer() widget is able to render texts dynamically according the data that is being provided.

Note to self: Look at how the Answer() widget is mapped.